smart_vision/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
Flask==2.2.2
fonttools==4.37.4
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
mahotas==1.4.13
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
opencv-contrib-python==4.7.0.72
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
PyJWT==2.5.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==0.23.0
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
twilio==7.14.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1